% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearchdomain_package.R
\docType{package}
\name{paws.cloudsearchdomain-package}
\alias{paws.cloudsearchdomain}
\alias{paws.cloudsearchdomain-package}
\title{paws.cloudsearchdomain: Amazon CloudSearch Domain}
\description{
You use the AmazonCloudSearch2013 API to upload
documents to a search domain and search those documents.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
